Changed "North Macedonia, Republic of" to "North Macedonia".
1. https://www.iso.org/obp/ui/#iso:code:3166:MK
2. https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#MK

The name of the country corresponding to the `MK` alpha2 code has been changed from `North Macedonia, Republic of` to `North Macedonia`.

The naming dispute for the previous name of the country has been settled. Like other countries in this list, the `title case` for the country is `North Macedonia`. The `the Republic of` is unnecessary.